requirements.txt
dotenv
langchain-groq
pydantic
exa-py
typing
fastapi
uvicorn
google-auth
google-api-python-client